move text and speech gen to separate store
