+++ web/store/index.ts
@@ -10,6 +10,7 @@ export * from './settings'
 export * from './memory'
 export * from './scenario'
 export * from './message'
+export * from './response'
 export * from './admin'
 export * from './presets'
 export * from './tag'
